feat: refactors core tests into module tests
